Id is long.

+++ src/com/darwinsys/todo/model/Task.java
@@ -2,10 +2,10 @@ package com.example.myaccount.model;
 
 public class ToDoItem {
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getName() {
@@ -21,7 +21,7 @@ public class ToDoItem {
 		this.description = description;
 	}
 	
-	int id;
+	long id;
 	String name;
 	String description;
 }
